requirements.txt
flask
flask_marshmallow
marshmallow-sqlalchemy
flask-sessionstore
flask_sqlalchemy
flask-apispec
gunicorn
flask_testing=0.7.1
nose=1.3.7
